Commit initial code
This has already undergone a few hours of work. It contains both initial proof of concept code for reading and writing XHTML files as well as the functional code for reading an actual Calibre-generated EPUB-formatted xhtml chapter files and writing them out with a couple basic transformations. It still has a long way to go, but I'm committing so I can work on this elsewhere. Like, this weekend, when I'm away from my Windows PC while in Conroe.