Place newly-generated TOC into xhtml subdirectory
I may regret this and undo it later. Not sure. For now, it's in line
with my new, overall approach of "deal at the root level and make very
strong assumptions about its directory structure." It's less "portable"
but easier to use since every epub-util's input and ouput takes place
at the root of an epub directory structure.