Divide transformer processing into $isRecipe and otherwise
Much of (well, *all* of) the process() algorithm is designed to work on
the main recipe XHTML files. They have a certain expected structure and
if that structure is not met, this falls apart pretty quickly.

To fix this, I made a check for isRecipe(), which is dumb: if it finds
two or more images in a file, then it's a recipe (because every recipe
has a QR code and a photo -- there was honestly no other XHTML tag or
attribute indicator that differentiated a "recipe" XHTML file).

For files not a recipe, I still had to extract the slug name to give
them uniform output file names. However, to that end, I don't like this.
I left myself a comment about how to fix this and I'm going to tackle
that bit next. (Mainly because non-recipe files still need to be
processed. e.g., changing `style.css` to `../css/manuscript.css`)