Add export concept
This used to be "toc" in that it would actually inject a file into the
external build's template. But it was the only thing that worked like
that and was getting a bit annoying. Instead, it writes it to the normal
output directory where everything else goes. AND there's a new export
target which sends the toc plus the css and all the images into the
output directory as well. This makes things cleaner and more uniform
and the mbk-cfo Makefile instead can just pull everything from one spot.

This was all part of my "let's do this over from scratch and see what
breaks" because I ran into so many "try agains" that I couldn't tell
how many things had real, permanent fixes and how many were untracked
or just one-offs.