Add new qr-linker utility
This utility replaces img tags with the "qr" id with an anchor tag that
then contains the original img tag. In short:

<img src="path" id="qr">

Becomes

<a href="url"><img src="path" id="qr"></a>

Thereby making the image (in this case, the QR code) clickable. This is
useful where the QR code is displayed on devices where it can be clicked
and does not need to be re-scanned (e.g., reading it on an iPhone).