Add support for new QR code and stats placement
Before, the stats were an H3 tag, just the same as ingredients, method,
and nutrition. But I never liked it. Figured out a way to put it right
under the title AND have it look okay on every platform after rendering.

The QR code never did have a good home, but it's now in the upper left
since the upper right was extremely problematic for portability.

The stats get moved in real time, during the main pass. The QR code
gets moved after the fact, in a post-processing pass. In hindsight, this
is also how I should have implemented qr-linker. But I'm way beyond that
level of petty refactoring at this point. I gotta get this done.