Change how isRecipe() is determined
Before, it would just guess. If a file had two or more images, then it
was a recipe. Now, the document is assumed to be a recipe. If it's not,
or if the normal isRecipe-processing is to be skipped, then the source
file now requires the following custom meta tag:

<meta name="mbk-spd2" data-page-type="nonrecipe"/>